+++ protected/humhub/modules/content/models/Content.php
@@ -12,8 +12,6 @@ use humhub\components\ActiveRecord;
 use humhub\components\behaviors\GUID;
 use humhub\components\behaviors\PolymorphicRelation;
 use humhub\components\Module;
-use humhub\libs\DbDateValidator;
-use humhub\modules\activity\helpers\ActivityHelper;
 use humhub\modules\admin\permissions\ManageUsers;
 use humhub\modules\content\components\ContentActiveRecord;
 use humhub\modules\content\components\ContentContainerActiveRecord;
@@ -26,6 +24,7 @@ use humhub\modules\content\live\NewContent;
 use humhub\modules\content\permissions\CreatePrivateContent;
 use humhub\modules\content\permissions\CreatePublicContent;
 use humhub\modules\content\permissions\ManageContent;
+use humhub\modules\content\services\ContentStateService;
 use humhub\modules\notification\models\Notification;
 use humhub\modules\search\libs\SearchHelper;
 use humhub\modules\space\models\Space;
@@ -236,17 +235,18 @@ class Content extends ActiveRecord implements Movable, ContentOwner, SoftDeletab
      */
     public function afterSave($insert, $changedAttributes)
     {
-        if (// New Content with State Published:
-            ($insert && $this->state == Content::STATE_PUBLISHED) ||
-            // Content Updated from Draft to Published
-            (array_key_exists('state', $changedAttributes) &&
-                $this->state == Content::STATE_PUBLISHED &&
-                $changedAttributes['state'] == Content::STATE_DRAFT
-            )) {
-            $this->processNewContent();
-        }
+        if (array_key_exists('state', $changedAttributes)) {
+            $model = $this->getPolymorphicRelation();
 
-        if ($insert || array_key_exists('state', $changedAttributes)) {
+            if ($this->getStateService()->isPublished()) {
+                // Run process for new content(Send notifications) only after publishing the Content
+                $this->processNewContent();
+                // Also run process for parent object in order to send notifications like mentioning users
+                if ($model instanceof ActiveRecord) {
+                    $model->afterSave($insert, $changedAttributes);
+                }
+            }
+
             $previousState = $changedAttributes['state'] ?? null;
             $this->trigger(self::EVENT_STATE_CHANGED, new ContentStateEvent([
                 'content' => $this,
@@ -254,13 +254,12 @@ class Content extends ActiveRecord implements Movable, ContentOwner, SoftDeletab
                 'previousState' => $previousState
             ]));
 
-            $model = $this->getPolymorphicRelation();
             if ($model instanceof ContentActiveRecord) {
                 $model->afterStateChange($this->state, $previousState);
             }
         }
 
-        if ($this->state === static::STATE_PUBLISHED) {
+        if ($this->getStateService()->isPublished()) {
             SearchHelper::queueUpdate($this->getModel());
         } else {
             SearchHelper::queueDelete($this->getModel());
@@ -394,9 +393,7 @@ class Content extends ActiveRecord implements Movable, ContentOwner, SoftDeletab
             'source_pk' => $this->getPrimaryKey(),
         ]);
 
-        $this->setState(self::STATE_DELETED);
-
-        if (!$this->save()) {
+        if (!$this->getStateService()->delete()) {
             return false;
         }
 
@@ -977,8 +974,8 @@ class Content extends ActiveRecord implements Movable, ContentOwner, SoftDeletab
             return $this->checkGuestAccess();
         }
 
-        // If content is draft, in trash, unapproved - restrict view access to editors
-        if ($this->state !== static::STATE_PUBLISHED) {
+        // If content is not published(draft, in trash, unapproved) - restrict view access to editors
+        if (!$this->getStateService()->isPublished()) {
             return $this->canEdit();
         }
 
@@ -1072,51 +1069,20 @@ class Content extends ActiveRecord implements Movable, ContentOwner, SoftDeletab
         return $this->created_at !== $this->updated_at && !empty($this->updated_at) && is_string($this->updated_at);
     }
 
-    public static function getAllowedStates(): array
+    public function getStateService(): ContentStateService
     {
-        return [
-            self::STATE_PUBLISHED,
-            self::STATE_DRAFT,
-            self::STATE_SCHEDULED,
-            self::STATE_DELETED
-        ];
-    }
-
-    /**
-     * @param int|string|null $state
-     * @return bool
-     * @since 1.14
-     */
-    public function canChangeState($state): bool
-    {
-        return in_array($state, self::getAllowedStates());
+        return new ContentStateService(['content' => $this]);
     }
 
     /**
      * @param int|string|null $state
      * @param array $options Additional options depending on state
      * @since 1.14
+     * @deprecated Use $this->getStateService()->set(). It will be deleted in v1.15.
      */
     public function setState($state, array $options = [])
     {
-        if (!$this->canChangeState($state)) {
-            return;
-        }
-
-        if ((int)$state === self::STATE_SCHEDULED) {
-            if (empty($options['scheduled_at'])) {
-                return;
-            }
-
-            $this->scheduled_at = $options['scheduled_at'];
-            (new DbDateValidator())->validateAttribute($this, 'scheduled_at');
-            if ($this->hasErrors('scheduled_at')) {
-                $this->scheduled_at = null;
-                return;
-            }
-        }
-
-        $this->state = $state;
+        $this->getStateService()->set($state, $options);
     }
 
 }